Moved more BSP task over to BSP-specific trait
