+++ core/api/src/mill/api/internal/api.scala
@@ -51,21 +51,10 @@ trait JavaModuleApi extends ModuleApi {
       needsToMergeResourcesIntoCompileDest: Boolean
   ): TaskApi[java.nio.file.Path]
 
-  @deprecated("Move to BSP context")
-  private[mill] def bspLoggingTest: TaskApi[Unit]
-
-  @deprecated("Move to BSP context")
   private[mill] def bspCompileClasspath(
       needsToMergeResourcesIntoCompileDest: Boolean
   )
       : TaskApi[EvaluatorApi => Seq[String]]
-
-  @deprecated("Move to BSP context")
-  private[mill] def bspBuildTargetScalacOptions(
-      needsToMergeResourcesIntoCompileDest: Boolean,
-      enableJvmCompileClasspathProvider: Boolean,
-      clientWantsSemanticDb: Boolean
-  ): TaskApi[(Seq[String], EvaluatorApi => Seq[String], EvaluatorApi => java.nio.file.Path)]
 
   private[mill] def genIdeaMetadata(
       ideaConfigVersion: Int,
@@ -89,20 +78,7 @@ trait JavaModuleApi extends ModuleApi {
 
 object JavaModuleApi
 
-trait BspJavaModuleApi extends ModuleApi {
 
-  private[mill] def bspRun(args: Seq[String]): TaskApi[Unit]
-
-  private[mill] def bspBuildTargetJavacOptions(
-      needsToMergeResourcesIntoCompileDest: Boolean,
-      clientWantsSemanticDb: Boolean
-  )
-      : TaskApi[EvaluatorApi => (
-          classesPath: Path,
-          javacOptions: Seq[String],
-          classpath: Seq[String]
-      )]
-}
 
 trait UnresolvedPathApi[P] {
   def resolve(outPath: P): P
